pyproject.toml
[tool.poetry]
name = "generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
typer = "^0.15.1"
faker = "^33.1.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
psycopg2-binary = "^2.9.10"
pg8000 = "^1.31.2"

[tool.poetry.scripts]
generator = "generator.cli.app:app"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
